requirements.txt
streamlit==1.43.2
altair==5.3.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
sentence-transformers==3.4.1
PyMuPDF==1.25.3
pdfplumber==0.11.5
fpdf==1.7.2
openai==1.35.3
matplotlib==3.8.4
transformers==4.41.2
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tiktoken==0.7.0
langchain==0.2.17
langchain-community==0.2.6
langchain-core==0.2.43
langchain-openai==0.1.9
langchain-text-splitters==0.2.4
langsmith==0.1.147
langchain-huggingface==0.0.3
python-dotenv==1.0.1
faiss-cpu==1.7.4
huggingface_hub==0.28.0
